fix(gradle-plugin): update to new plugin to fix publishing errors
This also allows us to drastically simplify the configuration of the gradle-plugin module.

**Automatic**
The `plugin-publish` plugin automatically applies all othe necessary plugins, so others were removed to allow it to on apply what is required.

**Shadow**
We removed the hacks to get it to publish `shadowJar` becuase that is now natively supported by the `plugin-publish` plugin.