Use pre-indexing to support resolved documents including overlays & extensions
Adds support for RAML overlays & extensions by creating a `ShapeIndex` that records all the necessary information that is lost during resolution.

Additionally, all resolution and lookups have been moved to the `ResolutionContext` itself and generators are required to use it instead of directly attempting these operations, ensuring the index is used when required.

Finally, the AMF and Kotln versions have been updated to the latest versions to ensure the vastly different new code targets the latest APIs.